llm_ensemble/requirements.txt
# LLM Ensemble Chess Engine — runtime dependencies

# Environment variable loading
python-dotenv>=1.0.0

# LLM SDKs
openai>=1.30.0               # OpenAI (GPT-4o) and Grok/DeepSeek (OpenAI-compatible)
anthropic>=0.28.0            # Anthropic Claude
google-generativeai>=0.7.0   # Google Gemini

# Testing
pytest>=7.0
